refactor: birthday 필드 타입을 변경하라 (+6 squashed commits) Squashed commits: [a6330b4] refactor: 사용하지 않는 메서드는 삭제하라 [5a347bf] refactor: dto validation을 변경하라 [9d2f19e] refactor: transactional적용 범위를 method로 변경하라 [0804d52] refactor: 사용하지 않는 dto클래스를 삭제하라 [574f791] refactor: 도메인 필드명을 변경하라 [12a9009] refactor: dto 패키지명을 변경하라
refactor: 도메인 필드명을 변경하라

refactor: 사용하지 않는 dto클래스를 삭제하라

refactor: transactional적용 범위를 method로 변경하라

refactor: dto validation을 변경하라

refactor: 사용하지 않는 메서드는 삭제하라

refactor: birthday 필드 타입을 변경하라
