+++ src/main/java/prgrms/neoike/controller/MemberController.java
@@ -3,12 +3,12 @@ package prgrms.neoike.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import prgrms.neoike.controller.dto.memberdto.MemberSaveRequest;
+import prgrms.neoike.controller.dto.member.MemberSaveRequest;
 import prgrms.neoike.controller.mapper.MemberMapper;
 import prgrms.neoike.service.MemberService;
 import prgrms.neoike.service.dto.drawticketdto.DrawTicketListResponse;
-import prgrms.neoike.service.dto.memberdto.MemberDto;
-import prgrms.neoike.service.dto.memberdto.MemberResponse;
+import prgrms.neoike.service.dto.member.MemberDto;
+import prgrms.neoike.service.dto.member.MemberResponse;
 
 import javax.validation.Valid;
 import java.net.URI;
