+++ src/main/java/prgrms/neoike/controller/dto/member/TokenResponse.java
@@ -1,0 +1,4 @@
+package prgrms.neoike.controller.dto.member;
+
+public record TokenResponse(String token){
+}
