--- src/main/java/prgrms/neoike/controller/dto/memberdto/MemberLoginRequest.java
@@ -1,12 +1,0 @@
-package prgrms.neoike.controller.dto.memberdto;
-
-import javax.validation.constraints.NotBlank;
-
-public record MemberLoginRequest(
-        @NotBlank
-        String email,
-
-        @NotBlank
-        String password
-) {
-}
