--- src/main/java/prgrms/neoike/controller/dto/memberdto/TokenResponse.java
@@ -1,4 +1,0 @@
-package prgrms.neoike.controller.dto.memberdto;
-
-public record TokenResponse(String token){
-}
