+++ src/main/java/prgrms/neoike/controller/mapper/MemberMapper.java
@@ -1,27 +1,20 @@
 package prgrms.neoike.controller.mapper;
 
-import prgrms.neoike.controller.dto.memberdto.MemberLoginRequest;
+import prgrms.neoike.controller.dto.member.MemberSaveRequest;
 import prgrms.neoike.domain.member.Address;
-import prgrms.neoike.service.dto.memberdto.LoginDto;
-import prgrms.neoike.service.dto.memberdto.MemberDto;
-import prgrms.neoike.controller.dto.memberdto.MemberSaveRequest;
+import prgrms.neoike.service.dto.member.MemberDto;
 
 public class MemberMapper {
 
     public static MemberDto toMemberDto(MemberSaveRequest memberSaveRequest) {
         return MemberDto.builder()
-                .name(memberSaveRequest.name())
-                .password(memberSaveRequest.password())
-                .phoneNumber(memberSaveRequest.phoneNumber())
-                .birthday(memberSaveRequest.birthday())
-                .email(memberSaveRequest.email())
-                .address(new Address(memberSaveRequest.city(), memberSaveRequest.street(), memberSaveRequest.zipcode()))
-                .gender(memberSaveRequest.gender())
-                .build();
+            .name(memberSaveRequest.name())
+            .password(memberSaveRequest.password())
+            .phoneNumber(memberSaveRequest.phoneNumber())
+            .birthday(memberSaveRequest.birthday())
+            .email(memberSaveRequest.email())
+            .address(new Address(memberSaveRequest.city(), memberSaveRequest.street(), memberSaveRequest.zipcode()))
+            .gender(memberSaveRequest.gender())
+            .build();
     }
-
-    public static LoginDto toLoginDto(MemberLoginRequest memberLoginRequest) {
-        return new LoginDto(memberLoginRequest.email(), memberLoginRequest.password());
-    }
-
 }
