+++ src/main/java/prgrms/neoike/service/CustomUserDetailService.java
@@ -32,6 +32,6 @@ public class CustomUserDetailService implements UserDetailsService {
     }
 
     private User createUser(Member member) {
-        return new User(member.getEmail().getEmail(), member.getPassword().getPassword(), List.of());
+        return new User(member.getEmail().getAddress(), member.getPassword().getPassword(), List.of());
     }
 }
