+++ src/main/java/prgrms/neoike/service/dto/member/LoginDto.java
@@ -1,4 +1,4 @@
-package prgrms.neoike.service.dto.memberdto;
+package prgrms.neoike.service.dto.member;
 
 public record LoginDto(String email, String password) {
 }
