+++ src/main/java/prgrms/neoike/service/dto/member/MemberDto.java
@@ -1,18 +1,18 @@
-package prgrms.neoike.service.dto.memberdto;
+package prgrms.neoike.service.dto.member;
 
 import lombok.Builder;
 import prgrms.neoike.domain.member.Address;
 import prgrms.neoike.domain.member.CountryType;
 import prgrms.neoike.domain.member.Gender;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 @Builder
 public record MemberDto(
         String email,
         String password,
         String name,
-        LocalDateTime birthday,
+        LocalDate birthday,
         Address address,
         CountryType countryCode,
         String phoneNumber,
